Reworked can.Fetch() functionality to improve error handling
Also removed the IgnitionState enum and put it in the DC code.
Deprecated error pointer (it was pointless). Now we store errors and
counters inside the class.

+++ v3.0/Layouts.cpp
@@ -176,7 +176,7 @@ Frame DC_Reset::generate_frame() const {
 
 Frame DC_SwitchPos::generate_frame() const {
 	Frame f;
-	f.s0 = static_cast<uint16_t>(state);
+	f.s0 = state;
 	f.s1 = UNUSED;
 	f.s2 = UNUSED;
 	f.s3 = UNUSED;
